Merge develop

+++ backend/src/main/java/com/jakubdeniziak/librarian/librarybook/service/LibraryBookDefaultService.java
@@ -8,7 +8,10 @@ import com.jakubdeniziak.librarian.librarybook.domain.LibraryBookTuple;
 import com.jakubdeniziak.librarian.librarybook.entity.LibraryBookKey;
 import com.jakubdeniziak.librarian.librarybook.mapper.LibraryBookMapper;
 import com.jakubdeniziak.librarian.librarybook.repository.LibraryBookJpaRepository;
+import com.jakubdeniziak.librarian.security.service.CallerService;
 import lombok.AllArgsConstructor;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -22,15 +25,20 @@ public class LibraryBookDefaultService implements LibraryBookService {
     private final LibraryBookMapper mapper;
     private final LibraryService libraryService;
     private final BookService bookService;
+    private final CallerService callerService;
 
     @Override
     public void save(LibraryBook libraryBook, UUID libraryId, UUID bookId) {
+        checkAuthorization(libraryId);
         LibraryBook initializedLibraryBook = getInitializedLibraryBook(libraryBook, libraryId, bookId);
         repository.save(mapper.map(initializedLibraryBook));
     }
 
     @Override
     public void saveAll(List<LibraryBookTuple> libraryBookTuples) {
+        libraryBookTuples.forEach(libraryBookTuple ->
+                checkAuthorization(libraryBookTuple.getLibraryId())
+        );
         List<LibraryBook> initializedLibraryBooks = libraryBookTuples.stream()
                 .map(libraryBookTuple -> getInitializedLibraryBook(libraryBookTuple.getLibraryBook(),
                         libraryBookTuple.getLibraryId(),
@@ -41,22 +49,26 @@ public class LibraryBookDefaultService implements LibraryBookService {
 
     @Override
     public LibraryBook find(UUID libraryId, UUID bookId) {
+        checkAuthorization(libraryId);
         return mapper.mapToDomain(repository.findById(createKey(libraryId, bookId))
                 .orElseThrow(ResourceNotFoundException::new));
     }
 
     @Override
     public List<LibraryBook> findAllByLibrary(UUID libraryId) {
+        checkAuthorization(libraryId);
         return mapper.mapToDomain(repository.findAllByLibrary_Id(libraryId));
     }
 
     @Override
+    @PreAuthorize("hasRole('ADMIN')")
     public List<LibraryBook> findAll() {
         return mapper.mapToDomain(repository.findAll());
     }
 
     @Override
     public void update(UUID libraryId, UUID bookId, LibraryBook updated) {
+        checkAuthorization(libraryId);
         LibraryBook libraryBook = find(libraryId, bookId);
         if (updated.getNumberOfCopies() != null) {
             libraryBook.setNumberOfCopies(updated.getNumberOfCopies());
@@ -72,6 +84,7 @@ public class LibraryBookDefaultService implements LibraryBookService {
 
     @Override
     public void delete(UUID libraryId, UUID bookId) {
+        checkAuthorization(libraryId);
         repository.deleteById(createKey(libraryId, bookId));
     }
 
@@ -83,6 +96,19 @@ public class LibraryBookDefaultService implements LibraryBookService {
         libraryBook.setLibrary(libraryService.find(libraryId));
         libraryBook.setBook(bookService.find(bookId));
         return libraryBook;
+    }
+
+    private void checkAuthorization(UUID libraryId) {
+        if (!isUserAuthorized(libraryId)) {
+            throw new AccessDeniedException("You are not authorized to delete this library");
+        }
+    }
+
+    private boolean isUserAuthorized(UUID libraryId) {
+        if (callerService.isCallerAdmin()) {
+            return true;
+        }
+        return libraryService.getOwnerId(libraryId).equals(callerService.getCallerId());
     }
 
 }
